+++ RELEASE_NOTES.md
@@ -4,7 +4,7 @@
 - Migration M2 repo → GitHub Packages. M2 repo is no longer supported. Last version: 1.16 and 1.17-SNAPSHOT
 - Add methods to domain-based classes: testHardcoded(), unknown()
 - Add/Update constants: TestsJunitConstants
-- Add utilities: SpringAuthoritiesUtility
+- Add utilities: SpringAuthoritiesUtility, TimestampUtility (methods: getStartOfMonthTimestamp)
 - Add postgres converters: Version
 - Add postgres persistable abstractions